+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # CHANGELOG
 
+## 2019/12/04
+
+Features:
+
+- Enable no proxy setting
+- Enable prefer ssh setting
+- Enable run command anywhere in a git directory
+- Apply setting on GUI
+
+Bug Fixed:
+
+- Fixed locate bug on Ubuntu 18.04 server
+
+Others:
+
+- [config.json](config/config.json) add 161 git platform
+
 ## 2019/11/29 `0.1.0`
 
 Features: